Patch by Byron Foster <byron_foster@yahoo.com>:
Here is a new patch that implements the java naming scheme
through attributes in the database schema that works with the new code
that you added.  the <database> element now has a new attribute
defaultNameConversion that may be either underscore, javaname,
or nochange.  The description of these are in the DTD.   Also the
<table> and <column> elements have an optional attribute
nameConversion that may also be one of the above.   Without any
method specification the behavior of the name conversion is as it was
before.


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@227737 13f79535-47bb-0310-9956-ffa450edef68