removed unneeded file
added java.io.Serializable to a bunch of classes/interfaces. the reason is
that i'm getting a NPE when Tomcat reloads the session and it is in the
Torque.getDatabaseMap (281) method. I have a feeling that the map is being
lost in the de-serialization or something...i dunno...it doesn't hurt.


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@227742 13f79535-47bb-0310-9956-ffa450edef68