Path by Byron Foster <bfoster@base2.cc>:
When the Id method is autoincrement and a Torque OM object is inserted
the primary key is not known until after the insert.  Once Torque
performs the insert it calls the setPrimaryKey method of the OM
object, since the primary key prior to the insert was probably null
the current ObjectKey and the one being set do not match so the
modified flag is set to true. So the quick tour:  after an insert is
performed the modified flag falsely indicates the object is modified.


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@227786 13f79535-47bb-0310-9956-ffa450edef68