patch by Byron Foster
    Here is a patch that saves a Torque OM object passed as a parameter
to the set method of  another Torque OM object.  Of course the second
object's schema table definition contains a foreign key reference to the
first OM object.

Here is a snippet of code that it creates:

   public void save(DBConnection dbCon) throws Exception
   {
           doPreSave();
       if (!alreadyInSave)
     {
       alreadyInSave = true;

       if (aQuestionType != null)
       {
           aQuestionType.save(dbCon);
           setQuestionType(aQuestionType);
       }

       if (aAccount != null)
       {
           aAccount.save(dbCon);
           setAccount(aAccount);
       }

       if (isModified())
       {
           if (isNew())
           {
               SectionPeer.doInsert((Section)this, dbCon);
               setNew(false);
           }
           else
           {
               SectionPeer.doUpdate((Section)this, dbCon);
           }
       }


    So, we first test if a given "set object" exists, if so save it,
then update the foreign key id reference property with a possible new Id
created by the save.  The set object is saved before the containing
object is saved since we need to get a foreign key id from a possibly
new set object.  When the containing object is saved we want to insure
that it contains a foreign key id for the set object.  If these object
are not modified, then nothing will be save because the set object id
will be compared to the current set object id and they will be found to
match.  In this case the modified flag is left as is, and no action is
taken.  The alreadyInSave flag prevents any infinite looping.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@227942 13f79535-47bb-0310-9956-ffa450edef68