patch by Nick Pomfret, modified by me, so that it produces compilable code.
The copy() method in the torque generated in the object model doesn't seem
to work as I would expect.

When using the copy() method on an object it correctly copies that object
but not those objects dependent upon it.  Looking at the automatically
generated code it would appear that copying dependent objects is what is
supposed to happen.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@228030 13f79535-47bb-0310-9956-ffa450edef68