Added patch to T2 from Edgar Gonzalez <egg@valhallaproject.com>:
The following patch is related to the treatement of
Criteria.isIgnoreCase() by BasePeer.createQueryString() and
BasePeer.createPreparedStatement().

In the current code (before patch) the Criteria.isIgnoreCase() value
ALWAYS overwrite the previous values of the
Criterions.isIgnoreCase(). So if you have a Criteria which
isIgnoreCase()==false and you add a Criterion which
isIgnoreCase()==true, when the BasePeer.createQueryString() or
BasePeer.createPreparedStatement() will be executed, the
Criterion.isIgnoreCase() will be false.

With this patch the Criteria.isIgnoreCase() value overwrite the
previous value of the Criterions.isIgnoreCase() ONLY when
Criteria.isIgnoreCase()==true.


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@228078 13f79535-47bb-0310-9956-ffa450edef68