From: thierry.lach@pentamark.com Reply-To: dev@scarab.tigris.org Date: Thu,  7 Mar 2002 08:12:24 -0800 (PST) To: dev@scarab.tigris.org Subject:   Oracle support
The Oracle schema is still not created correctly.

SCARAB_ATTACHMENTS column ATTACHMENT_DATA is defined in scarab-schema.xml
as having a type of VARBINARY.  Turbine maps this to a column type of RAW
for an Oracle database.  Oracle's definition of a RAW column requires a
length attached.  I had previously submitted a patch to build.xml
that replaced the RAW with LONG RAW.  This patch was removed at about the
same time that Turbine corrected defining BOOLEANINT as INTEGER(1).

Can we get a correction to this incorporated into Turbine?  I think
that it would be safe to map VARBINARY to the Oracle BLOB type.


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@228153 13f79535-47bb-0310-9956-ffa450edef68