Switched MethodResultCache to use groups, so that method results for the same method with different arguments can be invalidated as a group.
Removed hack in MethodCacheKey that was not working very well, and is
not needed due to the change in MethodResultCache.

Managers take on the task of invalidating caches.  The keeps the number
of listeners small and better behaved.  Since the manager handles the
cache it is not unreasonable for it to handle invalidation as well.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@228226 13f79535-47bb-0310-9956-ffa450edef68