patch from: Jeff Schnitzer <jeff@infohazard.org>
Vector becomes List/ArrayList, and Enumeration becomes Iterator.  After
this patch, Vector should not appear anywhere in the code (even in
comments) and Enumeration only appears as part of the Log4J API method
calls.  I checked each case to make sure that Vector synchronization is
not being relied upon (it never was).


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@228315 13f79535-47bb-0310-9956-ffa450edef68