patch by Will Glass-Husain <wglass@forio.com>
 templates/sql/id-table/Control.vm.
Changes the id_table_id in the id table to start at 1000 instead of 10.  Without
this, including more than 10 tables in the project schema causes the records in
the id_table to have overlapping primary keys between the project table records
and the turbine security table records.

 templates/sql/id-table/idtable.*  (changed for each database).
Changed the first id for each table to be 1000 instead of 100.  This is useful
if you are loading in initial data with project-datasql, with primary keys for
each record being set manually (not by the id broker).  Previously you were
limited to loading in less than 100 records, now you can load in up to 999.


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@228375 13f79535-47bb-0310-9956-ffa450edef68