added the ability to reference tables in a different schema.  The objects defined in the referencing schema are aware of the objects in the referenced schema.  But the object model in the referenced schema is unaffected.  Does not currently handle a circular relationship.
The tables in the related schemas should be in the same database.

Useful for addons/optional components that should not affect the core object
model.  Might also be useful as a way to overcome the TurbineUser/security
tables problem that has confounded turbine users for years.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@228416 13f79535-47bb-0310-9956-ffa450edef68