patch by Stephen Haberman <stephenh@chase3000.com>
In the generated save method, if an error occurs,
BasePeer.rollBackTransaction is called. However, if an error occurs in
the rollBackTransaction, the user loses anything about the original
error and only sees the rollBackEx.

Ignores the rollBackEx and just returns the
original exception from save


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@228418 13f79535-47bb-0310-9956-ffa450edef68