patch by 	Stephen Haberman <stephenh@chase3000.com>
I've attached a patch that lets you not specify managers in the
configuration and instead they default to the generated peers. It could
be considered a hack, but I think it works.

Basically in the managers template, I added DEFAULT_MANAGER_CLASSNAME,
which is a static string of the manager class itself. Then in
getManager(), instead of just calling Torque.getManager(key), I pass
that static string as the default classname.

Inside the Torque.getManager(key, defaultClassname), it checks to see if
the key is already in the managers map. If not, it adds the key with the
defaultClassname, just as if it had come from the configuration via the
initManager method.

So you can't use the straight Torque.getManager("ThisClass") from your
client code, but you can use ThisClassManager.getInstance() and it'll
work fine (and unless I'm using the managers wrong, I go with the
ThisClassManager.getInstance method anyway).

jmcnally - I modified this to synchronize the initManager method.  Not sure
that this does not still suffer MT problems, but if a new manager is
placed into the map, it would not cause problems with standard usage.

PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@228420 13f79535-47bb-0310-9956-ffa450edef68