PR: Obtained from: Submitted by: Reviewed by:
added table to the test schema to cause the error described by
Howard Lin here

http://www.mail-archive.com/turbine-torque-dev@jakarta.apache.org/msg00424.html

The problem is in the Table.getForeignKey(columnName) method.  This method
only gives the first fk the column is a part of.  In many cases this will
cause torque to miss additional functionality/relationships it could be adding.

In other cases it can give a fk that is not expected.  I have eliminated one
use in this category that was leading to torque objects with compile errors.
A complete fix requires a signature change to
getForeignKey(columnName, foreignTable).

In eliminating the use of the method in XmlToAppData, I went ahead and moved
the logic out of that class and into the model (Database to be exact) where
it belongs.


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@228610 13f79535-47bb-0310-9956-ffa450edef68