Merged the TORQUE_3_1_Branch into HEAD. The 3_1_Branch was tagged before merging with TORQUE_3_1_BRANCH_BEFORE_MERGE the HEAD was tagged before merging with TORQUE_HEAD_BEFORE_MERGE_OF_3_1_BRANCH
The basic idea was to take the generator from HEAD and the runtime and maven plugin from 3_1_BRANCH.
Because quite a few changes have amounted, I changed the name of the current version to 3.2

A detailed changelog can be found at the developers mailing list, the title of the mail is
changelog: merge of Torque_3_1_BRANCH with HEAD

I checked that the runtimetest runs against a Oracle 9.2 database.

Things which definitely need checking are
- check postgresql and Derby support
- The docs/changes.html page is not generated by "maven xdoc".
No idea why, but it is also not generated in the 3_1_Branch. They also do not appear on the website...

Although I tried to be very cautios, there are many changes and chances are that some bugs have sneaked in.
PLEASE TEST THIS !!!!


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@229664 13f79535-47bb-0310-9956-ffa450edef68