Avoid ConcurrentModificationException if data is inserted into the LargeSelect while the first page of data is already displayed. Thanks to Gareth Boden and Thomas Vandahl In Thomas Vandahl's words: LargeSelect is made to retrieve a list of objects in the background while the Velocity template already can display the first page of results. This is very handy for paging through long lists of objects.
The said Exception occurs if the background thread still adds objects to
the result list while Velocity already uses an Iterator on the sublist
of the first page. This usually happens when selects take a while.

The patch simply copies the page of results instead of returning a
sublist. This avoids the problem described above and just needs a little
bit more memory.

I have this working in a production environment for almost a year with
no problems whatsoever.

Submitted by: Gareth Boden


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@229670 13f79535-47bb-0310-9956-ffa450edef68