Added support for specifying database schema names at generate time, i.e. tablenames like "my_schema.my_table" can now be used in schema.xml
To do this, the java naming methods "underscore" and "javaname" in the generator were changed such that, in addition to removing underscores, they remove dots as well.
Also, a new java naming method called underscoreOmitSchema was introduced, which works like the underscore method but omits everything before the last dot.
As an example, the tablename "my_schema.my_TABLE" would produce the java name "MySchemaMyTable" with the method "underscore", "MySchemaMyTABLE" with the method "javaName", and "MyTable" with the method UnderscoreOmitSchema.

I did not know how to proceed with the "nochange" method, it will produce unvalid java names if schemata are used (java name "my_schema.my_TABLE").

Also some docs were added in the schema-howto.

The changes were tested for oracle, I will do tests for mysql and postgresql as well.


git-svn-id: http://svn.apache.org/repos/asf/db/torque/trunk@229675 13f79535-47bb-0310-9956-ffa450edef68