revert TORQUE-195 Remove autocommit
This removal of autocommit management done in version 4.0 was backported
unintentionally, as a whole new class was being backported. The second
reason for bringing the autocommit management back is that it DOES make
sense - when removed, one must either complicate the client code or the
deployment to various servers. Finally, this TransactionManager
implementation doesn't make much sense if it handles the transactions
only partially...